Use <paper-dropdown-menu> in future-compatible way
In its hybrid 2.0 version, <paper-dropdown-menu> will switch from distributing with <content> to using the new <slot>, which is a breaking change for users of the element. Specifically, instead of looking for the "dropdown-content" class, it will look for the slot of the same name. This commit prepares for that by making this code compatible with both.

Note that you don't need to update to Polymer 2.0 to use the 2.0 version of paper-dropdown-menu.