Update deprecated @apply and var() CSS syntax for Polymer 2.0
Polymer 2.0 is production ready, and we advise all projects to begin migration.

This change migrates uses of @apply to remove the deprecated parentheses syntax, and adds var() to use of CSS Custom Properties in the fallback position of a var() expression.

These changes are backwards-compatible with Polymer 1.0 applications, as well as browser-native implementations of CSS Custom Properties.